ai-agent/pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = ""
authors = ["cygkichi <9675041+cygkichi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
notebook = "^7.1.3"
python-dotenv = "^1.0.1"
langchain-anthropic = "^0.1.11"
langchain = "^0.1.16"
streamlit = "^1.33.0"
langchain-community = "^0.0.34"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
